fix: get nearest tech file
